refactor: plainToClass path
